+++ apps/web/pages/feed.tsx
@@ -3,6 +3,7 @@ import {
   Container,
 } from '@chakra-ui/react';
 
+
 export default function Home() {
   return (
     <Container maxW={'1128px'}>
