+++ packages/ui/components/data.tsx
@@ -1,0 +1,32 @@
+import {FeedItem} from "./Feed"
+
+  const FEED_ITEMS: Array<FeedItem> = [
+    {
+      id: "1",
+      name: 'Juan Pablo',
+      photo:'https://images.unsplash.com/photo-1520810627419-35e362c5dc07?ixlib=rb-1.2.1&q=80&fm=jpg&crop=faces&fit=crop&h=200&w=200&ixid=eyJhcHBfaWQiOjE3Nzg0fQ',
+      description: 'Actress, musician, songwriter and artist. PM for work inquires or #tag me in your posts',  
+      skills: ['React', 'Node', 'MongoDB'],
+        role: 'Full Stack Developer',
+    },
+    {
+      id: "2",  
+      name: 'Lucia Perez',
+        photo:'https://images.unsplash.com/photo-1520810627419-35e362c5dc07?ixlib=rb-1.2.1&q=80&fm=jpg&crop=faces&fit=crop&h=200&w=200&ixid=eyJhcHBfaWQiOjE3Nzg0fQ',
+        description: 'Actress, musician, songwriter and artist. PM for work inquires or #tag me in your posts',  
+        skills: ['React', 'Node', 'MongoDB'],
+          role: 'Full Stack Developer',
+      },
+      {
+        id: "3",
+        name: 'Pedro Alvarado',
+        photo:'https://images.unsplash.com/photo-1520810627419-35e362c5dc07?ixlib=rb-1.2.1&q=80&fm=jpg&crop=faces&fit=crop&h=200&w=200&ixid=eyJhcHBfaWQiOjE3Nzg0fQ',
+        description: 'Actress, musician, songwriter and artist. PM for work inquires or #tag me in your posts',  
+        skills: ['React', 'Node', 'MongoDB'],
+          role: 'Full Stack Developer',
+      },
+    ]
+
+export {
+    FEED_ITEMS
+}
